test: split autoservicio e2e tests by command
